# swagger/api-docs.json
{"swagger":"2.0","info":{"description":"AdaptNGO API documentation","version":"0.0.1","title":"AdaptNGO API","contact":{},"license":{}},"host":"localhost:9000","basePath":"/","tags":[{"name":"account-resource","description":"Account Resource"},{"name":"adapt-resource","description":"Adapt Resource"},{"name":"category-resource","description":"Category Resource"},{"name":"line-item-resource","description":"Line Item Resource"},{"name":"roles-resource","description":"Roles Resource"},{"name":"user-jwt-controller","description":"User JWT Controller"},{"name":"user-resource","description":"User Resource"}],"paths":{"/api/account":{"get":{"tags":["account-resource"],"summary":"getAccount","operationId":"getAccountUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserDTO"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["account-resource"],"summary":"saveAccount","operationId":"saveAccountUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"userDTO","description":"userDTO","required":true,"schema":{"$ref":"#/definitions/UserDTO"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/account/change-password":{"post":{"tags":["account-resource"],"summary":"changePassword","operationId":"changePasswordUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"passwordChangeDto","description":"passwordChangeDto","required":true,"schema":{"$ref":"#/definitions/PasswordChangeDTO"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/account/reset-password/finish":{"post":{"tags":["account-resource"],"summary":"finishPasswordReset","operationId":"finishPasswordResetUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"keyAndPassword","description":"keyAndPassword","required":true,"schema":{"$ref":"#/definitions/KeyAndPasswordVM"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/account/reset-password/init":{"post":{"tags":["account-resource"],"summary":"requestPasswordReset","operationId":"requestPasswordResetUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"mail","description":"mail","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/activate":{"get":{"tags":["account-resource"],"summary":"activateAccount","operationId":"activateAccountUsingGET","produces":["*/*"],"parameters":[{"name":"key","in":"query","description":"key","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/adapt/getBy/categories":{"get":{"tags":["adapt-resource"],"summary":"getAllLineItemsByCategory","operationId":"getAllLineItemsByCategoryUsingGET","produces":["*/*"],"parameters":[{"name":"category","in":"query","description":"category","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/LineItem"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/adapt/getBy/desc":{"get":{"tags":["adapt-resource"],"summary":"getAllLineItemsByDesc","operationId":"getAllLineItemsByDescUsingGET","produces":["*/*"],"parameters":[{"name":"desc","in":"query","description":"desc","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/LineItem"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/adapt/getBy/roles":{"get":{"tags":["adapt-resource"],"summary":"getAllLineItemsByRole","operationId":"getAllLineItemsByRoleUsingGET","produces":["*/*"],"parameters":[{"name":"role","in":"query","description":"role","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/LineItem"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/adapt/incrementViewCount/{id}":{"get":{"tags":["adapt-resource"],"summary":"addView","operationId":"addViewUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"boolean"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/authenticate":{"get":{"tags":["account-resource"],"summary":"isAuthenticated","operationId":"isAuthenticatedUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["user-jwt-controller"],"summary":"authorize","operationId":"authorizeUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"loginVM","description":"loginVM","required":true,"schema":{"$ref":"#/definitions/LoginVM"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/JWTToken"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/categories":{"get":{"tags":["category-resource"],"summary":"getAllCategories","operationId":"getAllCategoriesUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Category"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["category-resource"],"summary":"createCategory","operationId":"createCategoryUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"category","description":"category","required":true,"schema":{"$ref":"#/definitions/Category"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Category"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["category-resource"],"summary":"updateCategory","operationId":"updateCategoryUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"category","description":"category","required":true,"schema":{"$ref":"#/definitions/Category"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Category"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/categories/{id}":{"get":{"tags":["category-resource"],"summary":"getCategory","operationId":"getCategoryUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Category"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["category-resource"],"summary":"deleteCategory","operationId":"deleteCategoryUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/api/line-items":{"get":{"tags":["line-item-resource"],"summary":"getAllLineItems","operationId":"getAllLineItemsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/LineItem"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["line-item-resource"],"summary":"createLineItem","operationId":"createLineItemUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"lineItem","description":"lineItem","required":true,"schema":{"$ref":"#/definitions/LineItem"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/LineItem"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["line-item-resource"],"summary":"updateLineItem","operationId":"updateLineItemUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"lineItem","description":"lineItem","required":true,"schema":{"$ref":"#/definitions/LineItem"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/LineItem"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/line-items/incrementViews/{id}":{"get":{"tags":["line-item-resource"],"summary":"addView","operationId":"addViewUsingGET_1","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"boolean"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/line-items/{id}":{"get":{"tags":["line-item-resource"],"summary":"getLineItem","operationId":"getLineItemUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/LineItem"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["line-item-resource"],"summary":"deleteLineItem","operationId":"deleteLineItemUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/api/register":{"post":{"tags":["account-resource"],"summary":"registerAccount","operationId":"registerAccountUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"managedUserVM","description":"managedUserVM","required":true,"schema":{"$ref":"#/definitions/ManagedUserVM"}}],"responses":{"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/roles":{"get":{"tags":["roles-resource"],"summary":"getAllRoles","operationId":"getAllRolesUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Roles"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["roles-resource"],"summary":"createRoles","operationId":"createRolesUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"roles","description":"roles","required":true,"schema":{"$ref":"#/definitions/Roles"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Roles"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["roles-resource"],"summary":"updateRoles","operationId":"updateRolesUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"roles","description":"roles","required":true,"schema":{"$ref":"#/definitions/Roles"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Roles"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/roles/{id}":{"get":{"tags":["roles-resource"],"summary":"getRoles","operationId":"getRolesUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Roles"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["roles-resource"],"summary":"deleteRoles","operationId":"deleteRolesUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/api/users":{"get":{"tags":["user-resource"],"summary":"getAllUsers","operationId":"getAllUsersUsingGET","produces":["*/*"],"parameters":[{"name":"page","in":"query","description":"Page number of the requested page","required":false,"type":"integer","format":"int32"},{"name":"size","in":"query","description":"Size of a page","required":false,"type":"integer","format":"int32"},{"name":"sort","in":"query","description":"Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"type":"array","items":{"type":"string"},"collectionFormat":"multi"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/UserDTO"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["user-resource"],"summary":"createUser","operationId":"createUserUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"userDTO","description":"userDTO","required":true,"schema":{"$ref":"#/definitions/UserDTO"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/User"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["user-resource"],"summary":"updateUser","operationId":"updateUserUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"userDTO","description":"userDTO","required":true,"schema":{"$ref":"#/definitions/UserDTO"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserDTO"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/users/authorities":{"get":{"tags":["user-resource"],"summary":"getAuthorities","operationId":"getAuthoritiesUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"string"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/users/{login}":{"get":{"tags":["user-resource"],"summary":"getUser","operationId":"getUserUsingGET","produces":["*/*"],"parameters":[{"name":"login","in":"path","description":"login","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserDTO"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["user-resource"],"summary":"deleteUser","operationId":"deleteUserUsingDELETE","produces":["*/*"],"parameters":[{"name":"login","in":"path","description":"login","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}}},"definitions":{"Category":{"type":"object","properties":{"categoryName":{"type":"string"},"id":{"type":"integer","format":"int64"}},"title":"Category"},"JWTToken":{"type":"object","properties":{"id_token":{"type":"string"}},"title":"JWTToken"},"KeyAndPasswordVM":{"type":"object","properties":{"key":{"type":"string"},"newPassword":{"type":"string"}},"title":"KeyAndPasswordVM"},"LineItem":{"type":"object","properties":{"categories":{"type":"string"},"desc":{"type":"string"},"id":{"type":"integer","format":"int64"},"link":{"type":"string"},"name":{"type":"string"},"roles":{"type":"string"},"type":{"type":"string"},"viewCount":{"type":"integer","format":"int32"}},"title":"LineItem"},"LoginVM":{"type":"object","required":["password","username"],"properties":{"password":{"type":"string","minLength":4,"maxLength":100},"rememberMe":{"type":"boolean"},"username":{"type":"string","minLength":1,"maxLength":50}},"title":"LoginVM"},"ManagedUserVM":{"type":"object","properties":{"activated":{"type":"boolean"},"authorities":{"type":"array","items":{"type":"string"}},"createdBy":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"email":{"type":"string","minLength":5,"maxLength":254},"firstName":{"type":"string","minLength":0,"maxLength":50},"id":{"type":"integer","format":"int64"},"imageUrl":{"type":"string","minLength":0,"maxLength":256},"langKey":{"type":"string","minLength":2,"maxLength":10},"lastModifiedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"lastName":{"type":"string","minLength":0,"maxLength":50},"login":{"type":"string","minLength":1,"maxLength":50,"pattern":"^[_.@A-Za-z0-9-]*$"},"password":{"type":"string","minLength":4,"maxLength":100}},"title":"ManagedUserVM"},"PasswordChangeDTO":{"type":"object","properties":{"currentPassword":{"type":"string"},"newPassword":{"type":"string"}},"title":"PasswordChangeDTO"},"Roles":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"roleName":{"type":"string"}},"title":"Roles"},"User":{"type":"object","required":["activated","login"],"properties":{"activated":{"type":"boolean"},"email":{"type":"string","minLength":5,"maxLength":254},"firstName":{"type":"string","minLength":0,"maxLength":50},"id":{"type":"integer","format":"int64"},"imageUrl":{"type":"string","minLength":0,"maxLength":256},"langKey":{"type":"string","minLength":2,"maxLength":10},"lastName":{"type":"string","minLength":0,"maxLength":50},"login":{"type":"string","minLength":1,"maxLength":50,"pattern":"^[_.@A-Za-z0-9-]*$"},"resetDate":{"type":"string","format":"date-time"}},"title":"User"},"UserDTO":{"type":"object","properties":{"activated":{"type":"boolean"},"authorities":{"type":"array","items":{"type":"string"}},"createdBy":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"email":{"type":"string","minLength":5,"maxLength":254},"firstName":{"type":"string","minLength":0,"maxLength":50},"id":{"type":"integer","format":"int64"},"imageUrl":{"type":"string","minLength":0,"maxLength":256},"langKey":{"type":"string","minLength":2,"maxLength":10},"lastModifiedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"lastName":{"type":"string","minLength":0,"maxLength":50},"login":{"type":"string","minLength":1,"maxLength":50,"pattern":"^[_.@A-Za-z0-9-]*$"}},"title":"UserDTO"}}}